+++ CHANGELOG.md
@@ -36,6 +36,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Include appropriate trait bounds in return type of `AsyncMonitoredItem::into_stream()`.
 - Breaking: Add prefix `with_` in `ua::BrowseNextRequest::with_release_continuation_points()`.
 - Upgrade to open62541 released version 1.4.3.
+- Breaking: Change signatures of `AsyncClient::browse()` and `AsyncClient::browse_many()` to accept
+  `ua::BrowseDescription` instead of `ua::NodeId` for better control over the resulting references.
 
 ## [0.6.0-pre.5] - 2024-05-31